split diagnostics tab fixed log and crash dump directory links in label not working
