+++ src/settingsdialog.ui
@@ -1325,6 +1325,9 @@ programs you are intentionally running.</string>
          <property name="wordWrap">
           <bool>true</bool>
          </property>
+         <property name="openExternalLinks">
+          <bool>true</bool>
+         </property>
         </widget>
        </item>
        <item row="1" column="0">
